Split tests.
